GfPEPS.jl Documentation

GfPEPS.GaussianMap — Method

GaussianMap(CM_out::AbstractMatrix, CM_in::AbstractArray, Nf::Int, Λ::Int)

Returns the Gaussian map: CMout = B * inv(D + CMin) * B' + A. This contracts the virtual bonds and only the physical modes remain.

Keyword arguments:

CM_out::AbstractMatrix: The covariance matrix of the fiducial state / the covariance matrix dual to the Gaussian map

CM_in::AbstractMatrix: The covariance matrix of the virtual bonds

Nf::Int: number of physical fermions

Λ::Int: number of virtual fermions per bond

Note:

CMout must be a real antisymmetric matrix, i.e., CMout² = -I

The covariance matrices are currently only Fourier transformed TODO: also do for real space / no translation inv systems